VtolPathFollower: emergency land when gps data unuseable for navigation
Gives some predictable behavior to VTOL path follower in case of
GPS signal lost (or unusable quality). When GPS lost, it will wait
and continue operations for 2 seconds to ignore short-term GPS glitches
and then emergency landing is enabled. In emergency landing VTOL attitude
is stabilized to be level and throttle is controlled to achieve
configured landing  vertical speed.